+++ src/components/activities/Carousel.tsx
