+++ src/components/activities/CategoryMenu.tsx
