+++ src/components/activities/Layout.tsx
