+++ src/components/activities/type.d.ts
@@ -1,0 +1,16 @@
+import { Category } from '@trip.zip-api';
+
+export interface Activity {
+  id: number;
+  userId: number;
+  title: string;
+  description: string;
+  category: Category;
+  price: number;
+  address: string;
+  bannerImageUrl: string;
+  rating: number;
+  reviewCount: number;
+  createdAt: string;
+  updatedAt: string;
+}
